// src/types/errors.model.ts
export enum Errors {
    NOT_FOUND = "NOT_FOUND",
    ALREADY_EXISTS = "ALREADY_EXISTS",
    INCORRECT_PASSWORD = "INCORRECT_PASSWORD",
    UNPROCESSABLE_ENTITY = "UNPROCESSABLE_ENTITY",
    INTERNAL_EXCEPTION = "INTERNAL_EXCEPTION",
    UNAUTHORIZED = "UNAUTHORIZED",
    PRODUCT_ARRAY_EMPTY = "PRODUCT_ARRAY_EMPTY",
    CONFLICT = "CONFLICT"
}